backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.13.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]
pydantic[email]